requirements.txt
# Core
pandas
numpy
matplotlib
seaborn

# ML / NLP
scikit-learn
transformers
torch
sentence-transformers
nltk
spacy
textblob

# Utilities
requests
python-dotenv
tqdm

# Notebook Support
jupyter
ipykernel